refactor: split dataset tests in a tests mod (#5387)
The intention of this PR is to make our codebase a bit more AI-friendly
and developer-friendly. It's really hard to scroll through a large file
with 10k lines. This PR splits the dataset tests, over 7k LOC, into
smaller files to make them easier to find and edit.

I'm not sure if it's a good idea, I just put this together and am
looking for feedback.

---


**Parts of this PR were drafted with assistance from Codex (with
`gpt-5.1-codex-max`) and fully reviewed and edited by me. I take full
responsibility for all changes.**

---------

Signed-off-by: Xuanwo <github@xuanwo.io>