fix: dir namespace cloud storage path removes one subdir level (#5495)
We use `uri_to_url` to convert a root location to `Url` and then add
table subpath in dir namespace. This causes issue because there is no
trailing slash for the `uri`, and `Url::join` would treat the last part
of the path as a file, so joining it with another file would cause the
last part of the sub directory to be ignored.

This PR ensures that we add a trailing slash to the url. `uri_to_url` is
used all over the place to convert root paths, so far no place do `join`
after the conversion except for dir namespace, so there is no impact on
other places. But I think a more comprehensive fix would be to ensure a
trailing slash in the function directly, but it will be a fix that
affects more places so will do it later after examining all the use
cases more carefully.